+++ content-targeting-api/service/com/liferay/content/targeting/service/UserSegmentServiceUtil.java
@@ -95,13 +95,13 @@ public class UserSegmentServiceUtil {
 		return getService().getUserSegments(groupIds);
 	}
 
-	public static long getUserSegmentsCount(long groupId)
+	public static int getUserSegmentsCount(long groupId)
 		throws com.liferay.portal.kernel.exception.PortalException,
 			com.liferay.portal.kernel.exception.SystemException {
 		return getService().getUserSegmentsCount(groupId);
 	}
 
-	public static long getUserSegmentsCount(long[] groupIds)
+	public static int getUserSegmentsCount(long[] groupIds)
 		throws com.liferay.portal.kernel.exception.PortalException,
 			com.liferay.portal.kernel.exception.SystemException {
 		return getService().getUserSegmentsCount(groupIds);
